main scene - title etc.

+++ src/game/card.cpp
@@ -26,7 +26,7 @@ Card Deck::PopFrontCard()
 	return drawn_card;
 }
 
-int Deck::GetSize() { return cards.size(); }
+size_t Deck::GetSize() { return cards.size(); }
 
 const struct Card &Deck::GetConstCardReference(int index)
 {
